pyproject.toml
[project]
name = "GoWithMe"
version = "0.1.0"
description = "Event & Meetups Planner"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12"
package-mode = false
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "flask-wtf (>=1.2.2,<2.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-restx (>=1.3.2,<2.0.0)",
    "flask-login (>=0.6.3,<0.7.0)"
]



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"